backend/requirements.txt
aerich==0.7.2
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
bitarray==2.9.0
cached-property==1.5.2
caio==0.9.13
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cytoolz==0.12.2
dictdiffer==0.9.0
eth-abi==4.2.1
eth-account==0.10.0
eth-bloom==3.0.0
eth-hash==0.5.2
eth-keyfile==0.7.0
eth-keys==0.4.0
eth-rlp==1.0.0
eth-tester==0.9.1b1
eth-typing==3.5.2
eth-utils==2.3.1
fastapi==0.105.0
frozenlist==1.4.1
h11==0.14.0
hexbytes==0.3.1
idna==3.6
iso8601==1.1.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
lru-dict==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
parsimonious==0.9.0
protobuf==4.25.1
py-ecc==6.0.0
py-evm==0.7.0a4
py-geth==4.0.0
pycryptodome==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
pyethash==0.1.27
pypika-tortoise==0.1.6
python-dotenv==1.0.0
pytz==2023.3.post1
pyunormalize==15.1.0
referencing==0.32.0
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.15.2
safe-pysha3==1.0.4
semantic-version==2.10.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.27.0
tomlkit==0.12.3
toolz==0.12.0
tortoise-orm==0.20.0
trie==2.2.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
web3==6.12.0
websockets==12.0
yarl==1.9.4
